refactor para carga de textos con url y fixes varios
