+++ backend/.env.sample
@@ -9,11 +9,10 @@ DATABASE_HOST=127.0.0.1
 DATABASE_PORT=3307
 
 # AWS ENV configuration
-AWS_SAVE=false
 AWS_REGION=
 AWS_KEY=
 AWS_SECRET=
-AWS_BUCKET=
+AWS_BUCKET=limoneno
 
 #REDIS
 REDIS_URL=redis://localhost:16379/0
