+++ backend/app/services/aws_service.rb
@@ -29,9 +29,5 @@ module AwsService
         }
       })
     end
-
-    def self.getURL(object)
-      "https://#{object.bucket_name}.s3.amazonaws.com/#{object.key}"
-    end
   end
 end
