+++ backend/app/workers/text_extraction_worker.rb
@@ -1,0 +1,7 @@
+class TextExtractionWorker
+  include Sidekiq::Worker
+
+  def perform(id)
+    DatasetService::Files.extract_remote_text(id)
+  end
+end
