+++ backend/config/database.yml
@@ -9,7 +9,7 @@ default: &default
   encoding: utf8
   host: <%= ENV['DATABASE_HOST'] %>
   pool: <%= ENV.fetch("RAILS_MAX_THREADS") { 5 } %>
-  username: <%= ENV['DATABASE_USERNAME'] { 'root' } %>
+  username: <%= ENV.fetch('DATABASE_USERNAME') { 'root' } %>
   password: <%= ENV['DATABASE_PASSWORD'] %>
   database: <%= ENV['DATABASE_NAME'] %>
   port: <%= ENV['DATABASE_PORT'] %>
